// server/src/api.d.ts
export interface Booster {
	/**
	 * the booster ID
	 *
	 * boo_<set_id>-<booster_name>
	 */
	id: `boo_${string}-${string}`
	/**
	* the name of the booster
	*/
	name: string
	/**
	* the logo of the booster
	*/
	logo?: string
	/**
	* the front of the booster pack
	*/
	artwork_front?: string
	/**
	* the back of the booster pack
	*/
	artwork_back?: string
}

/**
 * /series endpoint
 */
export interface SerieResume {
	id: string;
	name: string;
	logo?: string;
}

/**
 * /series/:id endpoint
 */
export interface Serie extends SerieResume {
	firstSet: SetResume;
	lastSet: SetResume;
	releaseDate: string;
	sets: Array<SetResume>;
}

interface variants {
	normal?: boolean;
	reverse?: boolean;
	holo?: boolean;
	firstEdition?: boolean;
	wPromo?: boolean
}

interface variant_detailed {
	type: string
	size?: string
	stamp?: Array<string>
	foil?: string
}

export interface SetResume {
	id: string;
	name: string;
	logo?: string;
	symbol?: string;
	cardCount: {
		/**
		 * total of number of cards
		 */
		total: number;
		/**
		 * number of cards officialy (on the bottom of each cards)
		 */
		official: number;
	};
}

/**
 * /sets/:id
 */
export interface Set extends SetResume {
	serie: SerieResume;
	tcgOnline?: string;
	variants?: variants;
	releaseDate: string;

	/**
	 * The boosters available in this set
	 */
	boosters?: Array<Booster>
	/**
	 * Designate if the set is usable in tournaments
	 *
	 * Note: this is specific to the set and if a
	 * card is banned from the set it will still be true
	 */
	legal: {
		/**
		 * Ability to play in standard tournaments
		 */
		standard: boolean;
		/**
		 * Ability to play in expanded tournaments
		 */
		expanded: boolean;
	};
	cardCount: {
		/**
		 * total of number of cards
		 */
		total: number;
		/**
		 * number of cards officialy (on the bottom of each cards)
		 */
		official: number;
		/**
		 * number of cards having a normal version
		 */
		normal: number;
		/**
		 * number of cards having an reverse version
		 */
		reverse: number;
		/**
		 * number of cards having an holo version
		 */
		holo: number;
		/**
		 * Number of possible cards
		 */
		firstEd?: number;
	};
	cards: Array<CardResume>;
	abbreviation: { official: string, localized: string };
	thirdParty?: {
		cardmarket?: number
		tcgplayer?: number
	}
}
export interface CardResume {
	id: string;
	localId: string;
	/**
	 * Card Name (Including the suffix if next to card name)
	 */
	name: string;
	image?: string;
}